==> Chapter2/1.Trait_compilation/0.SOURCE_ALL.R <==
## Sources all scripts for the Trait compilation folder.

  source("1.1.Compile_trait_data.R")
  source("1.2.Compile_trait_data_without_taxonomic_correction.R")
  source("2.Filter_and_add_taxinfo.R")
  # source("3.Phylogenies_pseudoreplication.R") To be can be run independently. Results are not affected by upstream processes.
  source("4.Extract_phylogenetic_eigenvectors.R")
  source("5.Compare_with_without_taxonomic_corrections.R")
  source("6.Transform_Standardise.R")
